Enabling prescreening for ideation WIP
