+++ back/lib/participation_method/ideation.rb
@@ -15,7 +15,8 @@ module ParticipationMethod
     end
 
     def assign_defaults(input)
-      input_status_code = input.creation_phase&.prescreening_enabled ? 'prescreening' : 'proposed'
+      current_phase = TimelineService.new.current_phase(input.project)
+      input_status_code = current_phase&.prescreening_enabled ? 'prescreening' : 'proposed'
       input.idea_status ||= IdeaStatus.find_by!(code: input_status_code, participation_method: idea_status_method)
       input.publication_status ||= input.idea_status.public_post? ? 'published' : 'submitted'
     end
